feat: 로그인, 온보딩 API 작업 완료

+++ core/network/src/main/java/com/combo/runcombi/network/model/response/MemberModel.kt
@@ -6,12 +6,12 @@ import kotlinx.serialization.Serializable
 data class MemberModel(
     val email: String? = null,
     val gender: String? = null,
-    val height: Int? = null,
+    val height: Double? = null,
     val isActive: String? = null,
     val memberId: Int? = null,
     val memberTerms: List<String>? = null,
     val nickname: String? = null,
     val profileImgKey: String? = null,
     val profileImgUrl: String? = null,
-    val weight: Int? = null,
+    val weight: Double? = null,
 )
